+++ ebpf/src/kprobeperfevent.h
@@ -26,12 +26,12 @@ public:
   virtual Type type() const override;
 
   virtual const std::string &name() const override;
-  virtual std::uint64_t identifier() const override;
+  virtual std::uint32_t identifier() const override;
   virtual int fd() const override;
 
 protected:
   KprobePerfEvent(const std::string &name, bool ret_probe,
-                  std::uint64_t identifier, std::int32_t process_id);
+                  std::uint32_t identifier, std::int32_t process_id);
 
   friend class IPerfEvent;
 };
