+++ ebpf/src/tracepointperfevent.h
@@ -26,11 +26,11 @@ public:
   virtual Type type() const override;
 
   virtual const std::string &name() const override;
-  virtual std::uint64_t identifier() const override;
+  virtual std::uint32_t identifier() const override;
   virtual int fd() const override;
 
 protected:
-  TracepointPerfEvent(const std::string &name, std::uint64_t identifier,
+  TracepointPerfEvent(const std::string &name, std::uint32_t identifier,
                       std::int32_t process_id);
 
   friend class IPerfEvent;
