+++ ebpf/src/uprobeperfevent.h
@@ -26,12 +26,12 @@ public:
   virtual Type type() const override;
 
   virtual const std::string &name() const override;
-  virtual std::uint64_t identifier() const override;
+  virtual std::uint32_t identifier() const override;
   virtual int fd() const override;
 
 protected:
   UprobePerfEvent(const std::string &name, const std::string &path,
-                  bool ret_probe, std::uint64_t identifier,
+                  bool ret_probe, std::uint32_t identifier,
                   std::int32_t process_id);
 
   friend class IPerfEvent;
